+++ cloud/albuns/src/main/java/br/com/ada/albuns/model/dto/AlbumTemplateDTO.java
@@ -1,7 +1,6 @@
 package br.com.ada.albuns.model.dto;
 
 import java.math.BigDecimal;
-import java.sql.String;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -13,13 +12,13 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class AlbumTemplateDTO {
 
-  private Integer id;
+  private Long id;
 
   private String name;
 
   private String cover;
 
-  private Integer numStickers;
+  private Long numStickers;
 
   private BigDecimal price;
